Bluetooth: controller: LLL: Fix AuxPtr fill when for last PDU in chain
Last chain PDU should not have AuxPtr field.
It may happen that there are no other extended advertising
data in a PDU, so ext_hdr_len is zero. In such case there
should be no verification of aux_ptr field in the header and
the procedure of AuxPtr filling should be finished.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>