DNM: Bluetooth: controller: Make per. adv. sync PDU changing API global
Part of the API that is currently used only internally in
ull_aux_sync.c file will be required to be accessible
globally for use in ull_df.c.
The API should be refactored before that due to similiarity
of the code in particular functions. Also some functions
names should be updated due to similarity that may be
ambiguous and error prone.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>